refactor(utils): move dir
