+++ packages/bezier-react/src/components/Forms/Slider/Slider.tsx
@@ -8,7 +8,7 @@ import * as SliderPrimitive from '@radix-ui/react-slider'
 import {
   cssVarName,
   px,
-} from '~/src/utils/css'
+} from '~/src/utils/styleUtils'
 import { isNumber } from '~/src/utils/typeUtils'
 
 import {
