+++ packages/bezier-react/src/components/Tabs/TabList.tsx
@@ -6,7 +6,7 @@ import React, {
 import {
   cssVarName,
   px,
-} from '~/src/utils/css'
+} from '~/src/utils/styleUtils'
 
 import { TabListContextProvider } from './TabListContext'
 import {
